+++ src/pages/(admin)/register-student/index.tsx
@@ -19,11 +19,12 @@ export default function StudentRegistration() {
   const [password, setPassword] = useState("");
 
   const [isSubmitting, setIsSubmitting] = useState(false);
+  const [isSuggestingId, setIsSuggestingId] = useState(false);
 
   const handleRegistration = useCallback(
     async (e: SyntheticEvent) => {
       e.preventDefault();
-      setIsSubmitting(false);
+      setIsSubmitting(true);
       try {
         const response = await axios.post("/admins/register-users", {
           firstname,
@@ -39,6 +40,13 @@ export default function StudentRegistration() {
         // const modeSet = data?.mode_id
         // console.log({ response });
         dispatch(openSnackbar({ message: message, isError: false }));
+        setFirstname("");
+        setLastname("");
+        setEmail("");
+        setRegistrationId("");
+        setMatricNumber("");
+        setUsername("");
+        setPassword("");
       } catch (error: any) {
         const errorMessage = error?.response?.data?.message;
         dispatch(openSnackbar({ message: errorMessage, isError: true }));
@@ -56,6 +64,28 @@ export default function StudentRegistration() {
       registrationId,
     ]
   );
+
+  const fetchIdSuggestion = useCallback(async () => {
+    setIsSuggestingId(true);
+    try {
+      const response = await axios.get("/hardware/suggested-id");
+      const { data, message } = response?.data;
+      setRegistrationId(`${data?.previous_id}`);
+      dispatch(openSnackbar({ message: message, isError: false }));
+    } catch (error: any) {
+      const errorMessage = error?.response?.data?.message;
+      dispatch(
+        openSnackbar({
+          message: errorMessage
+            ? errorMessage
+            : "Ooop! Somethigng went wrong, please try again later",
+          isError: true,
+        })
+      );
+    } finally {
+      setIsSuggestingId(false);
+    }
+  }, []);
 
   return (
     <motion.div
@@ -83,7 +113,8 @@ export default function StudentRegistration() {
             </div>
 
             <ClickButtonMain
-              isLoading={isSubmitting}
+              clickHandler={fetchIdSuggestion}
+              isLoading={isSuggestingId}
               type="button"
               label="Fetch ID"
             />