+++ src/pages/(lecturer)/lecturer-dashboard/index.tsx
@@ -1,7 +1,10 @@
 import { motion } from "framer-motion";
 import LinkButtonMain from "../../../components/buttons/link_button_main";
+import { useAppSelector } from "../../../store/hooks";
 
 export default function LecturerDashboard() {
+  const { user } = useAppSelector((state) => state?.userAuthentication?.value);
+
   return (
     <motion.div
       initial={{ x: -100 }}
@@ -25,7 +28,7 @@ export default function LecturerDashboard() {
                   <h3 className=" text-2xl md:text-3xl font-semibold">
                     Lecturer's Dashboard
                   </h3>
-                  <p>lecturer@io.com</p>
+                  <p>{user?.email}</p>
                 </div>
                 <LinkButtonMain
                   url="/lecturer-course-registration"
@@ -49,11 +52,11 @@ export default function LecturerDashboard() {
                         <path d="M11.25 4.533A9.707 9.707 0 0 0 6 3a9.735 9.735 0 0 0-3.25.555.75.75 0 0 0-.5.707v14.25a.75.75 0 0 0 1 .707A8.237 8.237 0 0 1 6 18.75c1.995 0 3.823.707 5.25 1.886V4.533ZM12.75 20.636A8.214 8.214 0 0 1 18 18.75c.966 0 1.89.166 2.75.47a.75.75 0 0 0 1-.708V4.262a.75.75 0 0 0-.5-.707A9.735 9.735 0 0 0 18 3a9.707 9.707 0 0 0-5.25 1.533v16.103Z" />
                       </svg>
                     ),
-                    value: 3,
+                    value: 0,
                     subSection: [
                       {
                         id: 1.1,
-                        value: 2,
+                        value: 0,
                         label: "active",
                       },
                     ],
@@ -76,17 +79,17 @@ export default function LecturerDashboard() {
                         <path d="M5.082 14.254a8.287 8.287 0 0 0-1.308 5.135 9.687 9.687 0 0 1-1.764-.44l-.115-.04a.563.563 0 0 1-.373-.487l-.01-.121a3.75 3.75 0 0 1 3.57-4.047ZM20.226 19.389a8.287 8.287 0 0 0-1.308-5.135 3.75 3.75 0 0 1 3.57 4.047l-.01.121a.563.563 0 0 1-.373.486l-.115.04c-.567.2-1.156.349-1.764.441Z" />
                       </svg>
                     ),
-                    value: 200,
+                    value: 0,
                     subSection: [
                       {
                         id: 2.1,
-                        value: 100,
-                        label: "100 Level",
+                        value: 0,
+                        label: "Level",
                       },
                       {
                         id: 2.2,
-                        value: 50,
-                        label: "400 Level",
+                        value: 0,
+                        label: "Level",
                       },
                     ],
                   },
@@ -107,16 +110,16 @@ export default function LecturerDashboard() {
                         />
                       </svg>
                     ),
-                    value: 5,
+                    value: 0,
                     subSection: [
                       {
                         id: 3.1,
-                        value: 0.7,
+                        value: 0,
                         label: "average",
                       },
                       {
                         id: 3.2,
-                        value: 150,
+                        value: 0,
                         label: "total students",
                       },
                     ],