+++ CHANGELOG.md
@@ -1,5 +1,13 @@
+##### 05.31.2024
+
+> Admin student registeration flow completion
+> Admin, lecturer login api integration and UI development
+> Admin 'add-admin' and 'add-lecturer' UI development and API integration
+> Course registration csv parsing, logic redefinition and API integration
+
 ##### 05.26.2024
 
+> Course error fix testing
 > Authentication logic definition
 > Student's login api integration
 > Student's registration api integration